TOOLS-4100 Move the test connection-options helpers into common/testopts
`common/db`'s own tests cannot call `common/testutil`, because `testutil` imports `common/db`. They
worked around that by keeping copies of `testutil`'s auth/SSL/connection-string helpers in
`db_test.go`, and they could not call `testutil.GetToolOptions` at all — so they hardcode
`localhost:33333` and ignore `TOOLS_TESTING_MONGOD`. Fixing that needs a package both `testutil` and
`db` can import.

This commit is the move only; no behavior changes.

* `common/testutil/auth.go` -> `common/testopts/auth.go` - moved wholesale.
* `common/testutil/ssl_integration.go` -> `common/testopts/ssl_integration.go` - moved wholesale.
* `common/testutil/testutil.go` -> `common/testopts/testopts.go` - moved `GetToolOptions`,
  `GetBareArgs`, and the env var name, which is now exported as `testopts.URIEnvVar` because
  `testutil.SkipForAtlasCluster` still needs it.
* `common/db/db.go` -> `common/testopts/testopts.go` - moved `DefaultTestPort`. It is a test-only
  constant and `common/db` no longer has any use for it.
* All call sites updated; no forwarding wrappers were left behind.
