+++ common/db/db.go
@@ -53,11 +53,6 @@ const (
 	MaxBSONSize = 16 * 1024 * 1024 // 16MB - maximum BSON document size
 )
 
-// Default port for integration tests.
-const (
-	DefaultTestPort = "33333"
-)
-
 const (
 	// ignorable errors.
 	ErrDuplicateKeyCode         = 11000
