+++ integration/exportimport/query_test.go
@@ -12,7 +12,7 @@ import (
 	"time"
 
 	"github.com/mongodb/mongo-tools/common/options"
-	"github.com/mongodb/mongo-tools/common/testutil"
+	"github.com/mongodb/mongo-tools/common/testopts"
 	"github.com/mongodb/mongo-tools/mongoexport"
 	"github.com/mongodb/mongo-tools/mongoimport"
 	"go.mongodb.org/mongo-driver/v2/bson"
@@ -35,7 +35,7 @@ func (s *ExportImportSuite) TestRoundTripLimit() {
 	_, err := coll.InsertMany(s.Context(), docs)
 	s.Require().NoError(err)
 
-	exportToolOptions, err := testutil.GetToolOptions()
+	exportToolOptions, err := testopts.GetToolOptions()
 	s.Require().NoError(err)
 	exportToolOptions.Namespace = &options.Namespace{DB: dbName, Collection: collName}
 	me, err := mongoexport.New(mongoexport.Options{
@@ -57,7 +57,7 @@ func (s *ExportImportSuite) TestRoundTripLimit() {
 
 	s.Require().NoError(coll.Drop(s.Context()))
 
-	importToolOptions, err := testutil.GetToolOptions()
+	importToolOptions, err := testopts.GetToolOptions()
 	s.Require().NoError(err)
 	importToolOptions.Namespace = &options.Namespace{DB: dbName, Collection: collName}
 	mi, err := mongoimport.New(mongoimport.Options{
@@ -165,7 +165,7 @@ func (s *ExportImportSuite) TestRoundTripSortAndSkip() {
 	_, err := coll.InsertMany(s.Context(), docs)
 	s.Require().NoError(err)
 
-	exportToolOptions, err := testutil.GetToolOptions()
+	exportToolOptions, err := testopts.GetToolOptions()
 	s.Require().NoError(err)
 	exportToolOptions.Namespace = &options.Namespace{DB: dbName, Collection: collName}
 	me, err := mongoexport.New(mongoexport.Options{
@@ -186,7 +186,7 @@ func (s *ExportImportSuite) TestRoundTripSortAndSkip() {
 
 	s.Require().NoError(coll.Drop(s.Context()))
 
-	importToolOptions, err := testutil.GetToolOptions()
+	importToolOptions, err := testopts.GetToolOptions()
 	s.Require().NoError(err)
 	importToolOptions.Namespace = &options.Namespace{DB: dbName, Collection: collName}
 	mi, err := mongoimport.New(mongoimport.Options{
@@ -221,7 +221,7 @@ func (s *ExportImportSuite) exportAndImportWithQuery(
 		_, err := db.Collection("source").InsertMany(s.Context(), sourceDocs)
 		s.Require().NoError(err)
 	}
-	exportToolOptions, err := testutil.GetToolOptions()
+	exportToolOptions, err := testopts.GetToolOptions()
 	s.Require().NoError(err)
 	exportToolOptions.Namespace = &options.Namespace{DB: dbName, Collection: "source"}
 	me, err := mongoexport.New(mongoexport.Options{
@@ -239,7 +239,7 @@ func (s *ExportImportSuite) exportAndImportWithQuery(
 	s.Require().NoError(err)
 	s.Require().NoError(tmpFile.Close())
 
-	importToolOptions, err := testutil.GetToolOptions()
+	importToolOptions, err := testopts.GetToolOptions()
 	s.Require().NoError(err)
 	importToolOptions.Namespace = &options.Namespace{DB: dbName, Collection: "dest"}
 	mi, err := mongoimport.New(mongoimport.Options{
