+++ mongoexport/mongoexport_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/mongodb/mongo-tools/common/db"
 	"github.com/mongodb/mongo-tools/common/log"
 	"github.com/mongodb/mongo-tools/common/options"
+	"github.com/mongodb/mongo-tools/common/testopts"
 	"github.com/mongodb/mongo-tools/common/testtype"
 	"github.com/mongodb/mongo-tools/common/testutil"
 	"github.com/stretchr/testify/assert"
@@ -40,7 +41,7 @@ var (
 )
 
 func simpleMongoExportOpts() (Options, error) {
-	toolOptions, err := testutil.GetToolOptions()
+	toolOptions, err := testopts.GetToolOptions()
 	if err != nil {
 		return Options{}, fmt.Errorf(
 			"error getting tool options to create a mongoexport instance: %w",
@@ -349,7 +350,7 @@ func TestBrokenPipe(t *testing.T) {
 
 	args := append(
 		[]string{"run", filepath.Join("..", "mongoexport", "main")},
-		testutil.GetBareArgs()...,
+		testopts.GetBareArgs()...,
 	)
 	args = append(args, "--db", dbName, "--collection", collName)
 	testutil.AssertBrokenPipeHandled(t, exec.Command("go", args...))
@@ -362,7 +363,7 @@ func TestExportNamespaceValidation(t *testing.T) {
 	log.SetWriter(io.Discard)
 
 	for _, dbName := range []string{"test.bar", `test"bar`} {
-		toolOptions, err := testutil.GetToolOptions()
+		toolOptions, err := testopts.GetToolOptions()
 		require.NoError(t, err)
 		toolOptions.Namespace = &options.Namespace{DB: dbName, Collection: "foo"}
 		_, err = New(Options{
@@ -373,7 +374,7 @@ func TestExportNamespaceValidation(t *testing.T) {
 		assert.Error(t, err, "db name %q should be rejected", dbName)
 	}
 
-	toolOptions, err := testutil.GetToolOptions()
+	toolOptions, err := testopts.GetToolOptions()
 	require.NoError(t, err)
 	toolOptions.Namespace = &options.Namespace{DB: "test", Collection: "system.foobar"}
 	me, err := New(Options{
@@ -393,7 +394,7 @@ func TestExportNoData(t *testing.T) {
 	testtype.SkipUnlessTestType(t, testtype.IntegrationTestType)
 	log.SetWriter(io.Discard)
 
-	toolOptions, err := testutil.GetToolOptions()
+	toolOptions, err := testopts.GetToolOptions()
 	require.NoError(t, err)
 	toolOptions.Namespace = &options.Namespace{DB: "test", Collection: "mongoexport_no_data_test"}
 
@@ -437,7 +438,7 @@ func TestExportPretty(t *testing.T) {
 	})
 	require.NoError(t, err)
 
-	toolOptions, err := testutil.GetToolOptions()
+	toolOptions, err := testopts.GetToolOptions()
 	require.NoError(t, err)
 	toolOptions.Namespace = &options.Namespace{DB: dbName, Collection: collName}
 	me, err := New(Options{
@@ -485,7 +486,7 @@ func TestExportWritesToStdout(t *testing.T) {
 	require.NoError(t, err)
 
 	args := []string{"go", "run", filepath.Join("..", "mongoexport", "main")}
-	args = append(args, testutil.GetBareArgs()...)
+	args = append(args, testopts.GetBareArgs()...)
 	args = append(args, "--db", dbName, "--collection", collName)
 	cmd := exec.Command(args[0], args[1:]...)
 	stdout, err := cmd.Output()
@@ -574,7 +575,7 @@ func TestExportNestedFieldsCSV(t *testing.T) {
 	})
 	require.NoError(t, err)
 
-	toolOptions, err := testutil.GetToolOptions()
+	toolOptions, err := testopts.GetToolOptions()
 	require.NoError(t, err)
 	toolOptions.Namespace = &options.Namespace{DB: dbName, Collection: collName}
 
@@ -617,7 +618,7 @@ func TestExportNestedFieldsCSV(t *testing.T) {
 
 func newExportTestClient(t *testing.T, dbName string) *mongo.Client {
 	t.Helper()
-	toolOptions, err := testutil.GetToolOptions()
+	toolOptions, err := testopts.GetToolOptions()
 	require.NoError(t, err, "should get tool options")
 	sessionProvider, err := db.NewSessionProvider(*toolOptions)
 	require.NoError(t, err, "should create session provider")
@@ -631,7 +632,7 @@ func newExportTestClient(t *testing.T, dbName string) *mongo.Client {
 
 func exportWithType(t *testing.T, ns *options.Namespace, typeName string) (string, error) {
 	t.Helper()
-	toolOptions, err := testutil.GetToolOptions()
+	toolOptions, err := testopts.GetToolOptions()
 	require.NoError(t, err)
 	toolOptions.Namespace = ns
 	me, err := New(Options{
