+++ mongoimport/mongoimport_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/mongodb/mongo-tools/common"
 	"github.com/mongodb/mongo-tools/common/db"
 	"github.com/mongodb/mongo-tools/common/options"
+	"github.com/mongodb/mongo-tools/common/testopts"
 	"github.com/mongodb/mongo-tools/common/testtype"
 	"github.com/mongodb/mongo-tools/common/testutil"
 	. "github.com/smartystreets/goconvey/convey"
@@ -145,7 +146,7 @@ func countDocuments(t *testing.T, sessionProvider *db.SessionProvider) (int, err
 // getBasicToolOptions returns a test helper to instantiate the session provider
 // for calls to StreamDocument.
 func getBasicToolOptions() *options.ToolOptions {
-	toolOptions, err := testutil.GetToolOptions()
+	toolOptions, err := testopts.GetToolOptions()
 	if err != nil {
 		panic(fmt.Sprintf("could not get tool options: %v", err))
 	}
@@ -203,7 +204,7 @@ func NewMockMongoImport() *MongoImport {
 func getImportWithArgs(t *testing.T, additionalArgs ...string) (*MongoImport, error) {
 	t.Helper()
 
-	opts, err := ParseOptions(append(testutil.GetBareArgs(), additionalArgs...), "", "")
+	opts, err := ParseOptions(append(testopts.GetBareArgs(), additionalArgs...), "", "")
 	if err != nil {
 		return nil, fmt.Errorf("error parsing args: %v", err)
 	}
@@ -1635,7 +1636,7 @@ func TestImportBooleanType(t *testing.T) {
 	}
 	require.NoError(t, tmpFile.Close())
 
-	importToolOptions, err := testutil.GetToolOptions()
+	importToolOptions, err := testopts.GetToolOptions()
 	require.NoError(t, err)
 	importToolOptions.Namespace = &options.Namespace{DB: dbName, Collection: collName}
 	mi, err := New(Options{
@@ -1743,7 +1744,7 @@ func TestImportExtraFields(t *testing.T) {
 	fieldNames := []string{"a", "b", "c.xyz", "d.hij.lkm"}
 	fieldFilePath := writeFieldFile(t, tmpDir, "fieldfile", fieldNames)
 
-	toolOpts, err := testutil.GetToolOptions()
+	toolOpts, err := testopts.GetToolOptions()
 	require.NoError(t, err)
 	toolOpts.Namespace = &options.Namespace{DB: dbName, Collection: collName}
 	mi, err := New(Options{
@@ -1962,7 +1963,7 @@ func TestImportModeByID(t *testing.T) {
 
 func newImportTestClient(t *testing.T, dbName string) *mongo.Client {
 	t.Helper()
-	toolOptions, err := testutil.GetToolOptions()
+	toolOptions, err := testopts.GetToolOptions()
 	require.NoError(t, err, "should get tool options")
 	sessionProvider, err := db.NewSessionProvider(*toolOptions)
 	require.NoError(t, err, "should create session provider")
@@ -1984,7 +1985,7 @@ func importCollection(
 	ingestOpts IngestOptions,
 ) error {
 	t.Helper()
-	toolOptions, err := testutil.GetToolOptions()
+	toolOptions, err := testopts.GetToolOptions()
 	require.NoError(t, err)
 	toolOptions.Namespace = ns
 	mi, err := New(Options{
@@ -2002,7 +2003,7 @@ func importCollection(
 
 func importFromFile(t *testing.T, filePath, dbOverride, collOverride string) {
 	t.Helper()
-	toolOpts, err := testutil.GetToolOptions()
+	toolOpts, err := testopts.GetToolOptions()
 	require.NoError(t, err)
 	toolOpts.Namespace = &options.Namespace{DB: dbOverride, Collection: collOverride}
 	mi, err := New(Options{
@@ -2124,7 +2125,7 @@ func testImportFieldsForFormat(
 	})
 
 	t.Run(format+"/noFieldSpec", func(t *testing.T) {
-		toolOpts, err := testutil.GetToolOptions()
+		toolOpts, err := testopts.GetToolOptions()
 		require.NoError(t, err)
 		toolOpts.Namespace = &options.Namespace{DB: dbName, Collection: format + "testcoll"}
 		_, err = New(Options{
@@ -2150,7 +2151,7 @@ func importAndCheckFields(t *testing.T, client *mongo.Client, o importFieldsOpts
 	t.Helper()
 	collName := o.format + "testcoll"
 	require.NoError(t, client.Database(o.dbName).Collection(collName).Drop(t.Context()))
-	toolOpts, err := testutil.GetToolOptions()
+	toolOpts, err := testopts.GetToolOptions()
 	require.NoError(t, err)
 	toolOpts.Namespace = &options.Namespace{DB: o.dbName, Collection: collName}
 	var ffPtr *string
